fix: amount 의 타입을 BigDecimal로 수정

+++ src/main/java/com/github/kellyihyeon/stanceadmin/application/accountbook/dto/MembershipFeeForm.java
@@ -6,6 +6,7 @@ import com.github.kellyihyeon.stanceadmin.domain.member.MembershipFeeType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.Month;
 import java.util.List;
 
@@ -17,7 +18,7 @@ public class MembershipFeeForm {
 
     private final MembershipFeeType membershipFeeType;
 
-    private final Long amount;
+    private final BigDecimal amount;
 
     private final Month dueMonth;
